memlimit: explicit cast to (double)

+++ lib/util/memlimit.c
@@ -326,7 +326,7 @@ memtouse(size_t maxmem, double maxmemfrac, size_t * memlimit)
 	/* Only use the specified fraction of the available memory. */
 	if ((maxmemfrac > 0.5) || (maxmemfrac == 0.0))
 		maxmemfrac = 0.5;
-	memavail = (size_t)(maxmemfrac * memlimit_min);
+	memavail = (size_t)(maxmemfrac * (double)memlimit_min);
 
 	/* Don't use more than the specified maximum. */
 	if ((maxmem > 0) && (memavail > maxmem))
